+++ lib/shrine.rb
@@ -599,7 +599,9 @@ class Shrine
 
         # Delegates to `Shrine#default_url`.
         def default_url(**options)
-          store.default_url(options.merge(context))
+          url = store.default_url(options.merge(context))
+          warn "Overriding Shrine#default_url is deprecated and will be removed in Shrine 2. You should use the default_url plugin."
+          url
         end
 
         # The validation block provided by `Shrine.validate`.